Create Ways to increase the ratio ADR
Analysis on the drivers that influence the increase or decrease of the ADR ratio. First we start by explaining why ADR is important, then show attribute on data that impacts more on ADR and finally we end the paper with a disclose of recommendations on strategy in order to rise up the income.